Prepare for deployment.
