--- CC_Attribution-ShareAlike_4.0_International.xmp
@@ -1,42 +1,0 @@
-<?xpacket begin='' id=''?>
-<x:xmpmeta xmlns:x='adobe:ns:meta/'>
-  <rdf:RDF xmlns:rdf='http://www.w3.org/1999/02/22-rdf-syntax-ns#'>
-    <rdf:Description rdf:about=''
-		     xmlns:xapRights='http://ns.adobe.com/xap/1.0/rights/'>
-      <xapRights:Marked>True</xapRights:Marked>
-    </rdf:Description>
-    <rdf:Description rdf:about=''
-		     xmlns:xapRights='http://ns.adobe.com/xap/1.0/rights/'>
-      <xapRights:WebStatement rdf:resource='https://github.com/stoyicker/android-check-2'/>
-    </rdf:Description>
-    <rdf:Description rdf:about=''
-		     xmlns:xapRights='http://ns.adobe.com/xap/1.0/rights/'
-		     >
-      <xapRights:UsageTerms>
-	<rdf:Alt>
-	  <rdf:li xml:lang='x-default' >This work is licensed under a &lt;a rel=&#34;license&#34; href=&#34;http://creativecommons.org/licenses/by-sa/4.0/&#34;&gt;Creative Commons Attribution-ShareAlike 4.0 International License&lt;/a&gt;.</rdf:li>
-	  <rdf:li xml:lang='en' >This work is licensed under a &lt;a rel=&#34;license&#34; href=&#34;http://creativecommons.org/licenses/by-sa/4.0/&#34;&gt;Creative Commons Attribution-ShareAlike 4.0 International License&lt;/a&gt;.</rdf:li>
-	  </rdf:Alt>
-      </xapRights:UsageTerms>
-    </rdf:Description>
-    <rdf:Description rdf:about=''
-		     xmlns:dc='http://purl.org/dc/elements/1.1/'>
-      <dc:title>
-	<rdf:Alt>
-	  <rdf:li xml:lang='x-default'>android-check 2</rdf:li>
-	  <rdf:li xml:lang='en'>android-check 2</rdf:li>
-	</rdf:Alt>
-      </dc:title>
-    </rdf:Description>
-    <rdf:Description rdf:about=''
-      xmlns:cc='http://creativecommons.org/ns#'>
-      <cc:license rdf:resource='http://creativecommons.org/licenses/by-sa/4.0/'/>
-    </rdf:Description>
-    <rdf:Description rdf:about=''
-      xmlns:cc='http://creativecommons.org/ns#'>
-      <cc:attributionName>Jorge Antonio Diaz-Benito Soriano</cc:attributionName>
-    </rdf:Description>
-    
-  </rdf:RDF>
-</x:xmpmeta>
-<?xpacket end='r'?>
