+++ plugin/src/main/groovy/org/stoyicker/androidcheck/CommonCheck.groovy
@@ -1,36 +1,10 @@
-/*
- * Copyright (c) 2015 Noveo Group
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * Except as contained in this notice, the name(s) of the above copyright holders
- * shall not be used in advertising or otherwise to promote the sale, use or
- * other dealings in this Software without prior written authorization.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
+package org.stoyicker.androidcheck
 
-package com.noveogroup.android.check.common
-
-import com.noveogroup.android.check.CheckExtension
 import org.gradle.api.GradleException
 import org.gradle.api.Project
 
-import java.awt.Desktop
+import java.awt.*
+import java.util.List
 
 abstract class CommonCheck<Config extends CommonConfig> {
