+++ plugin/src/main/groovy/org/stoyicker/androidcheck/checkstyle/CheckstyleConfig.groovy
@@ -1,0 +1,10 @@
+package org.stoyicker.androidcheck.checkstyle
+
+import org.stoyicker.androidcheck.CommonConfig
+import org.gradle.api.Project
+
+class CheckstyleConfig extends CommonConfig {
+
+    CheckstyleConfig(Project project) { super(project) }
+
+}
