+++ plugin/src/main/groovy/org/stoyicker/androidcheck/findbugs/FindbugsConfig.groovy
@@ -1,0 +1,10 @@
+package org.stoyicker.androidcheck.findbugs
+
+import org.stoyicker.androidcheck.CommonConfig
+import org.gradle.api.Project
+
+class FindbugsConfig extends CommonConfig {
+
+    FindbugsConfig(Project project) { super(project) }
+
+}
