+++ plugin/src/main/groovy/org/stoyicker/androidcheck/pmd/PmdConfig.groovy
@@ -1,0 +1,10 @@
+package org.stoyicker.androidcheck.pmd
+
+import org.stoyicker.androidcheck.CommonConfig
+import org.gradle.api.Project
+
+class PmdConfig extends CommonConfig {
+
+    PmdConfig(Project project) { super(project) }
+
+}
