--- plugin/src/main/resources/META-INF/gradle-plugins/com.noveogroup.android.check.properties
@@ -1,1 +1,0 @@
-implementation-class=com.noveogroup.android.check.CheckPlugin
